Move TrialId into its file and refactor imports
