+++ .changeset/old-apricots-shave.md
@@ -2,4 +2,4 @@
 "functions-subscription": minor
 ---
 
-Add handler to create a trial
+[IOPLT-449] - Add handler to create a trial
