+++ apps/functions-subscription/src/domain/activation-request.ts
@@ -3,9 +3,10 @@ import { pipe } from 'fp-ts/lib/function';
 import * as TE from 'fp-ts/lib/TaskEither';
 import * as RTE from 'fp-ts/lib/ReaderTaskEither';
 import { NonEmptyString } from '@pagopa/ts-commons/lib/strings';
-import { TrialId, TrialIdCodec, UserIdCodec } from './subscription';
+import { UserIdCodec } from './subscription';
 import { Capabilities } from './capabilities';
 import { ItemAlreadyExists } from './errors';
+import { TrialId, TrialIdCodec } from './trial';
 
 // a unique brand for id of document with type request
 interface ActivationRequestIdBrand {
